requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
curl_cffi==0.11.1
fpdf==1.7.2
frozendict==2.4.6
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
multitasking==0.0.11
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDrive==1.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
soupsieve==2.7
stripe==12.2.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.61
